refactor: extract trident http feature into ferriskey-api-trident
