+++ libs/ferriskey-api-trident/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "ferriskey-api-trident"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-api-core = { path = "../ferriskey-api-core" }
+ferriskey-core = { path = "../../core" }
+axum = { workspace = true }
+axum-cookie = { workspace = true }
+axum-extra = { workspace = true }
+serde = { workspace = true }
+tracing = { workspace = true }
+url = { workspace = true }
+utoipa = { workspace = true }
+uuid = { workspace = true }
+validator = { workspace = true }
